refactor(RadioButtonPanel): クリック委譲ロジックをClickablePanelへ切り出しuse clientを削除
RadioButtonPanelはPanelへのクリックハンドラをJSXで組み立てるだけで、
そのハンドラ自体はサーバ・クライアント双方で評価可能な純粋な処理だった。
ハンドラをアタッチする箇所のみ`client/components/ClickablePanel.tsx`として
独立したクライアント境界に切り出すことで、RadioButtonPanel自体は
'use client'なしでServer/Client双方から利用できるようになる。

実Next.js buildで動作確認済み。RadioButtonPanelはclient reference
symbolを持たず、ClickablePanel/RadioButtonのみ独立クライアント参照として
遅延ロードされることを確認した。
