+++ packages/smarthr-ui/src/components/RadioButtonPanel/client/components/index.ts
@@ -1,0 +1,1 @@
+export { ClickablePanel } from './ClickablePanel'
